Update cjbbyte.c
Created different versions of two functions. I had a printCharBits() function, but I realized that I was going to need a function that actually returned a String. I am going to need that function down the road. While I was at it, I made the getnbytes() function to accompany the printnbytes() function. I also added an extra check to printnbytes().